requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-filebrowser==3.12.1
django-filebrowser-no-grappelli==3.8.0
django-grappelli==2.13.3
django-nested-inline==0.4.0
django-tinymce4-lite==1.7.5
idna==2.8
jsmin==2.2.2
Pillow==7.0.0
python-slugify==4.0.0
pytz==2019.3
requests==2.22.0
six==1.13.0
slugify==0.0.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
